% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{climatechange}
\alias{climatechange}
\title{Mid 21st Century Global Temperature Projection Data}
\format{
A data frame with 270 rows and 3 variables:
\describe{
  \item{\code{temp}}{[double] anomaly of 2069-2098 mean relative to
      the 1970-1999 mean}
  \item{\code{GCM}}{[factor] family of general circulation model provider}
  \item{\code{RCP}}{[double] representative concentration pathway}
 }
}
\source{
Extracted from package \code{bang}. The raw data from which the indices are calculated are monthly
  CMIP5 scenario runs for global surface air temperature from the KNMI Climate Explorer (\url{https://climexp.knmi.nl/})
  on 4/3/2015.
}
\usage{
climatechange
}
\description{
Indices of global temperature change from late 20th century (1970-1999)
to the end of the 21st century (2069-2098) based on data produced by the Fifth
Coupled Model Intercomparison Project (CMIP5).
}
\references{
Northrop, P.J. and B.D. Hall (2024) bang: Bayesian Analysis, No Gibbs, R package, version 1.0.4, \doi{10.32614/CRAN.package.bang}

Northrop, P.J. and R.E. Chandler (2014). Quantifying
  Sources of Uncertainty in Projections of Future Climate.
  \emph{Journal of Climate}, \strong{27}, 8793-8808.
  \doi{10.1175/JCLI-D-14-00265.1}

Van Vuuren, D. P., Edmonds, J., Kainuma, M., Riahi, K.
  Thomson, A., Hibbard, K., Hurtt, G. C., Kram, T., Krey, V.,
  Lamarque, J.-F. (2011). The representative concentration pathways:
  an overview. \emph{Climatic change}, \strong{109}, 5-31.
  \doi{10.1007/s10584-011-0148-z}
}
\keyword{datasets}
