% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_helpers.R
\name{crw_Q}
\alias{crw_Q}
\title{Sparse circulant precision matrix for random walk priors on integers}
\usage{
crw_Q(d, type = c("rw1", "rw2"), sparse = TRUE)
}
\arguments{
\item{d}{integer; dimension of the model.}

\item{type}{type of prior, either first-order (\code{rw1}) or second-order (\code{rw2})}

\item{sparse}{logical; if \code{TRUE}, returns a sparse matrix of class \code{dsCMatrix}}
}
\value{
a \code{d} by \code{d} precision matrix
}
\description{
Given a dimension \code{d}, return a circulant matrix with the coefficients
}
