% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{efl}
\alias{efl}
\title{English Football League results for 2023-2024}
\format{
A data frame with 760 rows and 4 variables:
\describe{
  \item{\code{score}}{[integer] number of goals of \code{team} during a match}
  \item{\code{team}}{[factor] name of the team which scored the goal}
  \item{\code{opponent}}{[factor] name of adversary}
  \item{\code{home}}{[integer] binary variable worth 1 if \code{team} is playing at home, 0 otherwise.}
}
}
\usage{
efl
}
\description{
Scores of football (soccer) regular matches for the 2023 season of the EFL. Each of the 20 teams played twice against each opponent at home and away (four matches in total against each opponent).
}
\keyword{datasets}
