% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{equicorrelation}
\alias{equicorrelation}
\title{Equicorrelation matrix}
\usage{
equicorrelation(d, rho = 0, precision = FALSE)
}
\arguments{
\item{d}{dimension of the matrix}

\item{rho}{correlation between pairs}

\item{precision}{logical; if \code{TRUE}, returns the precision matrix, otherwise the correlation matrix (default to \code{FALSE})}
}
\value{
a \code{d} by \code{d} square matrix
}
\description{
Given a dimension and a correlation, returns the covariance or precision matrix
of the compound symmetry model.
}
