% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exchangerate}
\alias{exchangerate}
\title{Euro-American dollar exchange rate}
\format{
A data frame with 6335 rows and 2 variables:
\describe{
  \item{\code{date}}{[Date] date of measurement}
  \item{\code{dexrate}}{[double] exchange rate EUR/USD}
}
}
\source{
Board of Governors of the Federal Reserve System (US), U.S. Dollars to Euro Spot Exchange Rate [DEXUSEU], retrieved from FRED, Federal Reserve Bank of St. Louis; \url{https://fred.stlouisfed.org/series/DEXUSEU}, February 17, 2025.
}
\usage{
exchangerate
}
\description{
Irregular time series of official U.S. Dollars to Euro Spot Exchange Rate.
Week-ends and official holidays are excluded. The time series runs from 2015-02-07 until 2025-02-07.
}
\keyword{datasets}
