% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logist_KS}
\alias{logist_KS}
\title{Conditional Kolmogorov-Smirnov distribution draw}
\usage{
logist_KS(r)
}
\arguments{
\item{r}{vector of absolute value of centered augmented data}
}
\value{
vector of scale parameters \eqn{\lambda}
}
\description{
Follows the logistic regression, generates from the conditional
distribution of \eqn{\kappa} and \eqn{\lambda} given
the value of \eqn{R = |y - X\beta|} outlined in Appendix A.4 of Holmes and Held (2006)
}
\references{
Leonhard Held, Chris C. Holmes (2006). \emph{Bayesian auxiliary variable models for binary and multinomial regression}, Bayesian Analysis, 1(\bold{1}), 145-168.
}
