% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_helpers.R
\name{mh}
\alias{mh}
\title{Update for Markov chain Monte Carlo}
\usage{
mh(
  type = c("rw", "mala", "laplace"),
  par_curr,
  par_name,
  loglik,
  loglik_grad = NULL,
  loglik_hessian = NULL,
  logprior,
  logprior_grad = NULL,
  logprior_hessian = NULL,
  lb = -Inf,
  ub = Inf,
  transform = FALSE,
  damping = 1,
  sd_prop,
  ...
)
}
\arguments{
\item{type}{[string] update type, either random walk (\code{rw}), adjusted Langevin (\code{mala}) or quadratic approximation (\code{laplace})}

\item{par_curr}{[double] current value of the parameter}

\item{par_name}{[string] the name of the argument for the function}

\item{loglik}{[function] log likelihood function}

\item{loglik_grad}{[function] derivative of the log likelihood with respect to parameter of interest}

\item{loglik_hessian}{[function] second derivative of the log likelihood with respect to the parameter of interest}

\item{logprior}{[function] log prior of parameter}

\item{logprior_grad}{[function] derivative of the log prior with respect to parameter of interest}

\item{logprior_hessian}{[function] second derivative of the log prior with respect to the parameter of interest}

\item{lb}{[scalar] lower bound for the parameter}

\item{ub}{[scalar] upper bound for the parameter}

\item{transform}{[logical] should the parameter be mapped onto the real line (\code{TRUE}) or should proposals be made based on truncated distributions (\code{FALSE})}

\item{damping}{[scalar] contraction factor for the Newton update}

\item{sd_prop}{[double] standard deviation of the proposal}

\item{...}{additional arguments passed to the log likelihood function and its derivatives}
}
\value{
value of the parameter
}
\description{
This function computes a single update for a Metropolis-Hastings
algorithm for univariate proposal based on random walk, adjusted Langevin
dynamics with gradient or a quadratic Laplace approximation.
}
