% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mtlprecip}
\alias{mtlprecip}
\title{Montreal precipitation data}
\format{
A data frame with 1827 rows and 3 variables:
\describe{
  \item{\code{year}}{[integer] year of observation}
  \item{\code{day}}{[integer] day of the year}
  \item{\code{precip}}{[double] amount of precipitation (in mm)}
}
}
\source{
Environment Canada
}
\usage{
mtlprecip
}
\description{
Amount of precipitation (in mm) in Montreal YUL airport
weather station for 2020-2024, by day of the year
}
\keyword{datasets}
