% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_helpers.R
\name{multivariate_laplace_update}
\alias{multivariate_laplace_update}
\title{Multivariate updates based on Laplace approximation}
\usage{
multivariate_laplace_update(
  par_curr,
  par_name,
  loglik,
  loglik_grad = NULL,
  loglik_hessian = NULL,
  logprior,
  logprior_grad = NULL,
  logprior_hessian = NULL,
  lb = -Inf,
  ub = Inf,
  damping = 1,
  ...
)
}
\arguments{
\item{par_name}{[string] the name of the argument for the function}

\item{loglik}{[function] log likelihood function}

\item{loglik_hessian}{[function] hessian matrix, the second derivative of the log likelihood with respect to the parameter of interest}

\item{logprior}{[function] log prior of parameter}

\item{logprior_hessian}{[function] second derivative of the log prior with respect to the parameter of interest}

\item{lb}{[double] vector of lower bounds for the parameters}

\item{ub}{[double] vector of upper bounds for the parameters}

\item{damping}{[double] double or vector of contraction factor for the Newton update}

\item{...}{additional arguments passed to the log likelihood function and its derivatives}

\item{cur}{[double] current value of the vector parameter}

\item{loglik_gradient}{[function] derivative of the log likelihood with respect to parameter of interest}

\item{logprior_gradient}{[function] derivative of the log prior with respect to parameter of interest}
}
\value{
a new vector for the parameter of interest
}
\description{
Following Rue and Held (2004), we perform
a random walk based on a multivariate normal approximation
at the \code{cur} parameter value.
}
\references{
Rue, H. and L. Held (2005), Gaussian Markov random fields, CRC, section 4.4.2
}
\keyword{internal}
