% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{posterior_smartwatch}
\alias{posterior_smartwatch}
\title{Posterior draws for Poisson mixed model}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 250 rows and 36 columns.
}
\usage{
posterior_smartwatch
}
\description{
A matrix containing 250 draws from the posterior of the 36 parameters for the Poisson mixed
model applied to the smartwatch dataset.
}
\keyword{internal}
