% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rats}
\alias{rats}
\title{Rat tumor data}
\format{
A data frame with 70 rows and 2 variables:
\describe{
  \item{\code{counts}}{[integer] counts of tumors out of \code{n}}
  \item{\code{n}}{[integer] experiment sample size}
}
}
\usage{
rats
}
\description{
Data from Tarone (1982), extracted from Table 5. It contains observations from
70 experiments on animal carcinogenesis bioassay conducted with female F344 rats,
with data on sample sizes and records of the number of lung cancer tumors developed.
}
\references{
Tarone, R. E. (1982) The use of historical information in
  testing for a trend in proportions. \emph{Biometrics}, \strong{38},
  215-220. \doi{10.2307/2530304}
}
\keyword{datasets}
