% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{rcondmvnorm}
\alias{rcondmvnorm}
\title{Simulation from a conditional multivariate Gaussian}
\usage{
rcondmvnorm(n, value, ind, mean, covariance, precision = NULL)
}
\arguments{
\item{n}{sample size}

\item{ind}{d vector containing the index of variables to simulate}

\item{mean}{vector of means}

\item{covariance}{covariance matrix}

\item{precision}{precision matrix}
}
\value{
an \code{n} by \code{d} matrix of draws
}
\description{
Simulation from a conditional multivariate Gaussian
}
