% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_helpers.R
\name{rGaussQ}
\alias{rGaussQ}
\title{Simulate a Gaussian random field}
\usage{
rGaussQ(n, b, Q)
}
\arguments{
\item{n}{integer; sample size}

\item{b}{\code{d} vector of scaled location}

\item{Q}{\code{d} by \code{d} precision matrix}
}
\value{
an \code{n} by \code{d} matrix
}
\description{
Simulate a Gaussian random field with sparse precision matrix \eqn{Q} and mean \eqn{Q^{-1}b}, exploiting the sparsity structure.
}
\references{
Rue, H. (2001). \emph{Fast sampling of Gaussian Markov random fields}. Journal of the Royal Statistical Society: Series B (Statistical Methodology), 63: 325-338. \doi{10.1111/1467-9868.00288}.
}
