% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_helpers.R
\name{rtnorm}
\alias{rtnorm}
\alias{dtnorm}
\title{Univariate truncated normal sampler}
\usage{
rtnorm(n = 1, mean = 0, sd = 1, a = -Inf, b = Inf)

dtnorm(x, mean = 0, sd = 1, a = -Inf, b = Inf, log = FALSE)
}
\arguments{
\item{n}{[integer] sample size}

\item{mean}{[numeric] scalar location parameter}

\item{sd}{[numeric] scalar scale parameter}

\item{a}{[numeric] lower bound}

\item{b}{[numeric] upper bound}

\item{x}{[numeric] observation vector}

\item{log}{[logical] if \code{TRUE}, return the log density}
}
\description{
This implementations works for a < 37
or b > 37
}
