% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_helpers.R
\name{transfo}
\alias{transfo}
\title{Transform parameter to unconstrained scale}
\usage{
transfo(par, lb, ub)
}
\arguments{
\item{par}{scalar parameter}

\item{lb}{lower bound}

\item{ub}{upper bound}
}
\description{
Transform parameter to unconstrained scale
}
\references{
Section 56 of the Stan Reference manual version 2.9 at \url{https://github.com/stan-dev/stan/releases/download/v2.9.0/stan-reference-2.9.0.pdf}
}
