% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_helpers.R
\name{univariate_laplace_update}
\alias{univariate_laplace_update}
\title{Univariate updates based on Laplace approximation}
\usage{
univariate_laplace_update(
  par_curr,
  par_name,
  loglik,
  loglik_grad = NULL,
  loglik_hessian = NULL,
  logprior,
  logprior_grad = NULL,
  logprior_hessian = NULL,
  lb = -Inf,
  ub = Inf,
  damping = 1,
  ...
)
}
\arguments{
\item{par_curr}{[double] current value of the parameter}

\item{par_name}{[string] the name of the argument for the function}

\item{loglik}{[function] log likelihood function}

\item{loglik_grad}{[function] derivative of the log likelihood with respect to parameter of interest}

\item{loglik_hessian}{[function] second derivative of the log likelihood with respect to the parameter of interest}

\item{logprior}{[function] log prior of parameter}

\item{logprior_grad}{[function] derivative of the log prior with respect to parameter of interest}

\item{logprior_hessian}{[function] second derivative of the log prior with respect to the parameter of interest}

\item{lb}{[scalar] lower bound for the parameter}

\item{ub}{[scalar] upper bound for the parameter}

\item{damping}{[scalar] contraction factor for the Newton update}

\item{...}{additional arguments passed to the log likelihood function and its derivatives}
}
\value{
a new value for the parameter of interest
}
\description{
Following Rue and Held (2004), we perform
a random walk based on a normal approximation
at the \code{cur} parameter value. This requires both
the gradient and the hessian of the log-likelihood and log-posterior;
if the latter are left undefined, they are obtained using numerical
differentiation via the \code{numDeriv} package.
In such cases, users cannot have a formal argument of the functions matching those of
\code{grad} and \code{hessian} (typically \code{x} or \code{method})
}
\details{
The algorithm uses a random walk proposal
and a Metropolis acceptance step.
}
\references{
Rue, H. and L. Held (2005), Gaussian Markov random fields, CRC press, section 4.4.1
}
\keyword{internal}
