% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_helpers.R
\name{univariate_mala_update}
\alias{univariate_mala_update}
\title{Univariate updates based on MALA}
\usage{
univariate_mala_update(
  par_curr,
  par_name,
  loglik,
  loglik_grad = NULL,
  logprior,
  logprior_grad = NULL,
  sd_prop,
  lb = -Inf,
  ub = Inf,
  damping = 1,
  transform = FALSE,
  ...
)
}
\description{
Univariate updates based on MALA
}
\keyword{internal}
