% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_helpers.R
\name{univariate_rw_update}
\alias{univariate_rw_update}
\title{Univariate updates based on random walk}
\usage{
univariate_rw_update(
  par_curr,
  par_name,
  loglik,
  logprior,
  sd_prop,
  lb = -Inf,
  ub = Inf,
  transform = FALSE,
  ...
)
}
\arguments{
\item{par_curr}{[double] current value of the parameter}

\item{par_name}{[string] the name of the argument for the function}

\item{loglik}{[function] log likelihood function}

\item{logprior}{[function] log prior of parameter}

\item{lb}{[scalar] lower bound for the parameter}

\item{ub}{[scalar] upper bound for the parameter}

\item{...}{additional arguments passed to the log likelihood function and its derivatives}
}
\value{
a new value for the parameter of interest
}
\description{
The algorithm uses a random walk proposal.
}
\keyword{internal}
