% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{upworthy_question}
\alias{upworthy_question}
\title{Upworthy headlines with questions}
\format{
A data frame with 10590 rows and 4 variables:
\describe{
  \item{\code{id}}{[factor] dummy identifier for questions}
  \item{\code{question}}{[factor] does headline contain a question, either \code{"yes"} or \code{"no"}}
  \item{\code{impressions}}{[integer] number of views}
  \item{\code{clicks}}{[integer] number of clicks}
}
}
\usage{
upworthy_question
}
\description{
Aggregated data from the Upworthy archive. Each observation represents the counts and impressions for a story that contained headlines with both questions or not, detected by searching for a question mark character. Only stories with both are preserved, so data are balanced
}
\keyword{datasets}
