% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{upworthy_sesame}
\alias{upworthy_sesame}
\title{Sesame street story headline}
\format{
A data frame with 4 rows and 3 variables:
\describe{
  \item{\code{headline}}{[factor] headline of story, one of \code{H1}, \code{H2}, \code{H3} or \code{H4}}
  \item{\code{impressions}}{[integer] number of views}
  \item{\code{clicks}}{[integer] number of clicks}
}
}
\usage{
upworthy_sesame
}
\description{
Data for impressions on different headlines for a story on
Sesame street by Upworthy (id 546f9c5c92f391a45e00003e).
}
\keyword{datasets}
