% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uscitypopn}
\alias{uscitypopn}
\title{Population of largest American cities}
\format{
A data frame with 113 rows and 2 variables:
\describe{
  \item{\code{city}}{[character] name of city}
  \item{\code{population}}{[integer] number of inhabitants}
}
}
\source{
Wikipedia, List of United States Cities by population. Accessed October 10th, 2023. \url{https://en.wikipedia.org/wiki/List_of_United_States_cities_by_population}
}
\usage{
uscitypopn
}
\description{
Size of cities above 200 000 inhabitants according to 2020 US census.
}
\keyword{datasets}
