% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exchangerate}
\alias{exchangerate}
\title{Euro-American dollar exchange rate}
\format{
A data frame with 6335 rows and 2 variables:
\describe{
  \item{\code{date}}{[Date] date of measurement}
  \item{\code{rate}}{[double] exchange rate EUR/USD}
}
}
\source{
European Central Bank’s Statistical Data Warehouse, accessed September 25th, 2023
}
\usage{
exchangerate
}
\description{
Irregular time series of official exchange rate
from the European Central Bank to American dollar from 1999 until April 2023.
Week-ends and official holidays are excluded.
}
\keyword{datasets}
